--HG--
branch : new-ui

+++ modules/forum/viewtopic.php
@@ -29,9 +29,11 @@ require_once 'config.php';
 require_once 'functions.php';
 require_once 'include/lib/modalboxhelper.class.php';
 require_once 'include/lib/multimediahelper.class.php';
+require_once 'include/course_settings.php';
 require_once 'modules/search/indexer.class.php';
 require_once 'modules/search/forumtopicindexer.class.php';
 require_once 'modules/search/forumpostindexer.class.php';
+require_once 'modules/rating/class.rating.php';
 
 ModalBoxHelper::loadModalBox();
 
@@ -47,15 +49,30 @@ if (isset($_GET['all'])) {
 
 if (isset($_GET['forum'])) {
     $forum = intval($_GET['forum']);
+} else {
+    header("Location: index.php?course=$course_code");
+    exit();
 }
 if (isset($_GET['topic'])) {
     $topic = intval($_GET['topic']);
 }
-$myrow = Database::get()->querySingle("SELECT f.id, f.name FROM forum f, forum_topic t
+if (isset($_GET['post_id'])) {//needed to find post page for anchors
+    $post_id = intval($_GET['post_id']);
+    $myrow = Database::get()->querySingle("SELECT f.id, f.name, p.post_time, p.poster_id, t.locked FROM forum f, forum_topic t, forum_post p
+            WHERE f.id = ?d
+            AND t.id = ?d
+            AND p.id = ?d
+            AND t.forum_id = f.id
+            AND p.topic_id = t.id
+            AND f.course_id = ?d", $forum, $topic, $post_id, $course_id);
+} else {
+    $myrow = Database::get()->querySingle("SELECT f.id, f.name, t.locked FROM forum f, forum_topic t
             WHERE f.id = ?d
             AND t.id = ?d
             AND t.forum_id = f.id
             AND f.course_id = ?d", $forum, $topic, $course_id);
+}
+
 
 if (!$myrow) {
     $tool_content .= "<p class='alert1'>$langErrorTopicSelect</p>";
@@ -64,26 +81,36 @@ if (!$myrow) {
 }
 $forum_name = $myrow->name;
 $forum = $myrow->id;
+$topic_locked = $myrow->locked;
 $total = get_total_posts($topic);
 
-if (isset($_GET['delete']) && $is_editor) {
+if (isset($_GET['delete']) && isset($post_id) && $is_editor) {
     $idx = new Indexer();
     $ftdx = new ForumTopicIndexer($idx);
     $fpdx = new ForumPostIndexer($idx);
 
-    $post_id = intval($_GET['post_id']);
     $last_post_in_thread = get_last_post($topic);
 
-    $myrow = Database::get()->querySingle("SELECT post_time FROM forum_post
-                            WHERE id = ?d", $post_id);
-
     $this_post_time = $myrow->post_time;
+    $this_post_author = $myrow->poster_id;
 
     Database::get()->query("DELETE FROM forum_post WHERE id = ?d", $post_id);
     $fpdx->remove($post_id);
+    
+    //orphan replies get -1 to parent_post_id
+    Database::get()->query("UPDATE forum_post SET parent_post_id = -1 WHERE parent_post_id = ?d", $post_id);
 
+    $forum_user_stats = Database::get()->querySingle("SELECT COUNT(*) as c FROM forum_post
+                        INNER JOIN forum_topic ON forum_post.topic_id = forum_topic.id
+                        INNER JOIN forum ON forum.id = forum_topic.forum_id
+                        WHERE forum_post.poster_id = ?d AND forum.course_id = ?d", $this_post_author, $course_id);
+    Database::get()->query("DELETE FROM forum_user_stats WHERE user_id = ?d AND course_id = ?d", $this_post_author, $course_id);
+    if ($forum_user_stats->c != 0) {
+        Database::get()->query("INSERT INTO forum_user_stats (user_id, num_posts, course_id) VALUES (?d,?d,?d)", $this_post_author, $forum_user_stats->c, $course_id);
+    }
+    
     if ($total == 1) { // if exists one post in topic
-        Database::get()->query("DELETE FROM forum_topic WHERE id = ?d AND forum_id = ", $topic, $forum);
+        Database::get()->query("DELETE FROM forum_topic WHERE id = ?d AND forum_id = ?d", $topic, $forum);
         $ftdx->remove($topic);
         Database::get()->query("UPDATE forum SET num_topics = 0,
                             num_posts = 0
@@ -135,11 +162,16 @@ if (isset($_SESSION['message'])) {
     $tool_content .= $_SESSION['message'];
     unset($_SESSION['message']);
 }
-$tool_content .= "<div id='operations_container'>
-	<ul id='opslist'>
-	<li><a href='reply.php?course=$course_code&amp;topic=$topic&amp;forum=$forum'>$langReply";
 
-$tool_content .= "</a></li></ul></div>";
+if ($topic_locked == 1) {
+    $tool_content .= "<p class='alert1'>$langErrorTopicLocked</p>";
+} else {
+    $tool_content .= "<div id='operations_container'>
+    	<ul id='opslist'>
+    	<li><a href='reply.php?course=$course_code&amp;topic=$topic&amp;forum=$forum'>$langReply";
+    
+    $tool_content .= "</a></li></ul></div>";
+}
 
 if ($paging and $total > $posts_per_page) {
     $times = 1;
@@ -150,6 +182,12 @@ if ($paging and $total > $posts_per_page) {
 	  <span class='row'><strong class='pagination'>
 	  <span>";
 
+    if (isset($post_id)) {
+        $result = Database::get()->querySingle("SELECT COUNT(*) as c FROM forum_post WHERE topic_id = ?d AND post_time <= ?t", $topic, $myrow->post_time);
+        $num = $result->c;
+        $_GET['start'] = (ceil($num/$posts_per_page)-1)*$posts_per_page;
+    }
+    
     if (isset($_GET['start'])) {
         $start = intval($_GET['start']);
     } else {
@@ -228,13 +266,20 @@ if (isset($_GET['all'])) {
 }
 
 $count = 0;
+$user_stats = array();
 foreach ($result as $myrow) {
     if ($count % 2 == 1) {
         $tool_content .= "<tr class='odd'>";
     } else {
         $tool_content .= "<tr class='even'>";
     }
-    $tool_content .= "<td valign='top'>" . display_user($myrow->poster_id) . "</td>";
+    
+    if (!isset($user_stats[$myrow->poster_id])) {
+        $user_num_posts = Database::get()->querySingle("SELECT num_posts FROM forum_user_stats WHERE user_id = ?d AND course_id = ?d", $myrow->poster_id, $course_id);
+        $user_stats[$myrow->poster_id] = $user_num_posts->num_posts;
+    }
+    
+    $tool_content .= "<td valign='top'>" . display_user($myrow->poster_id) . "<br/>".$user_stats[$myrow->poster_id]." $langMessages</td>";
     $message = $myrow->post_text;
     // support for math symbols
     $message = mathfilter($message, 12, "../../courses/mathimg/");
@@ -243,11 +288,29 @@ foreach ($result as $myrow) {
     } else {
         $postTitle = "";
     }
+    
+    $rate_str = "";
+    if (setting_get(SETTING_FORUM_RATING_ENABLE, $course_id)) {
+        $rating = new Rating('thumbs_up', 'forum_post', $myrow->id);
+        $rate_str = $rating->put($is_editor, $uid, $course_code);
+    }
+    
+    $anchor_link = "<a href='$_SERVER[SCRIPT_NAME]?course=$course_code&amp;topic=$topic&amp;forum=$forum&amp;post_id=$myrow->id#$myrow->id'>#$myrow->id</a><br/>"; 
+    if ($myrow->parent_post_id == -1) {
+        $parent_post_link = "<br/><br/>$langForumPostParentDel";
+    } elseif ($myrow->parent_post_id != 0) {
+        $parent_post_link = "<br/><br/>$langForumPostParent<a href='viewtopic.php?course=$course_code&amp;topic=$topic&amp;forum=$forum&amp;post_id=$myrow->parent_post_id#$myrow->parent_post_id'>#$myrow->parent_post_id</a>";
+    } else {
+        $parent_post_link = "";
+    }
+    
     $tool_content .= "<td>
 	  <div>
+	    <a name='".$myrow->id."'></a>$anchor_link
+	    <a href='reply.php?course=$course_code&amp;topic=$topic&amp;forum=$forum&amp;parent_post=$myrow->id'>$langForumPostReply</a><br/>
 	    <b>$langSent: </b>" . $myrow->post_time . "<br>$postTitle
 	  </div>
-	  <br />$message<br />
+	  <br />$message<br />".$rate_str.$parent_post_link."
 	</td>";
     if ($is_editor) {
         $tool_content .= "<td width='40' valign='top'>
